+++ client/src/common/protocol.ts
@@ -67,9 +67,3 @@ export const ReadFileRequest = new RequestType<string, number[], void>(
 export const ReadDirectoryRequest = new RequestType<string, string[], void>(
   "elm/readDirectory",
 );
-
-export const ProvideFileContentsRequest = new RequestType<
-  { uri: string },
-  string,
-  void
->("elm/provideFileContents");